Removed tabs from TabBarView and added them to a new file called TabBarTabs
